# testing-service/docs/swagger.yaml
definitions:
...
  ErrorMessage:
    properties:
      errorDate:
        description: Time of the error, formatted as yyyy-mm-dd hh:mm:ss.mmm.
        example: "2024-05-17 11:33:23.038"
        type: string
      errorMessage:
        description: The error itself, which may be a custom message or an internal
          failure.
        example: Something went wrong.
        type: string
      serviceName:
        description: Name of the service that produced the message.
        example: testing-service
        type: string
      stacktrace:
        description: The stack trace, when one is available; otherwise "No Stacktrace
          Available".
        example: No Stacktrace Available
        type: string
    type: object
  Filter:
    properties:
      condition:
        type: string
      feature:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
...
  SelectionSpecification:
    properties:
      filters:
        items:
          $ref: '#/definitions/Filter'
        type: array
      ids:
        items:
          type: string
        type: array
      searchText:
        type: string
    type: object
...
  TestCaseRunView:
    properties:
      chainId:
        type: string
      errors:
        type: integer
      finish:
        type: string
      id:
        type: string
      ordinal:
        type: integer
      sessionId:
        type: string
      start:
        type: string
      status:
        type: string
      testCaseDescription:
        type: string
      testCaseId:
        type: string
      testCaseName:
        type: string
      testsRunId:
        type: string
    type: object
...
info:
  contact: {}
  description: API of the testing service of the Qubership Integration Platform.
  title: Testing Service API
paths:
...
  /api/v1/test-case-runs:
    post:
      consumes:
      - application/json
      operationId: findTestCaseRunsV1
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Sort field
        enum:
        - id
        - test_case_name
        - chain_id
        - start
        - finish
        - status
        - errors
        in: query
        name: sort_by
        type: string
      - default: ASC
        description: Sort order
        enum:
        - ASC
        - DESC
        in: query
        name: sort_order
        type: string
      - description: Return IDs list
        in: query
        name: return_ids
        type: boolean
      - description: Test case runs specification
        in: body
        name: specification
        required: true
        schema:
          $ref: '#/definitions/SelectionSpecification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/TestCaseRunView'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Get test case runs
      tags:
      - V1
      - Test Case Runs
  /api/v1/test-case-runs/{id}:
    get:
      operationId: findTestCaseRunByIdV1
      parameters:
      - description: Test case run UUID
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TestCaseRunView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Get test case run
      tags:
      - V1
      - Test Case Runs
  /api/v1/test-case-runs/{id}/cancel:
    post:
      operationId: cancelTestCaseRunByIdV1
      parameters:
      - description: Test case run UUID
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Cancel test case run
      tags:
      - V1
      - Test Case Runs
...
  /api/v1/test-case-runs/{id}/export:
    post:
      consumes:
      - application/json
      operationId: exportTestCaseRunV1
      parameters:
      - description: Test case run UUID
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Export tests case run
      tags:
      - V1
      - Test Case Runs
  /api/v1/test-case-runs/cancel:
    post:
      consumes:
      - application/json
      operationId: cancelTestCaseRunsV1
      parameters:
      - description: Test case runs IDs to cancel
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Cancel test case runs
      tags:
      - V1
      - Test Case Runs
...
  /api/v1/test-case-runs/export:
    post:
      consumes:
      - application/json
      operationId: exportTestCaseRunsV1
      parameters:
      - description: Test case runs IDs to export
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - text/csv
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Export test case runs
      tags:
      - V1
      - Test Case Runs
...
produces:
- application/json
swagger: "2.0"
...